requirements.txt
google-genai>=1.50.0
pydantic>=2.0.0
pypdf>=6.0.0
python-dotenv>=1.0.0
